Create seekg.11.18.23.16:44
In C++, the seekg() function is used to set the position of the next character to be extracted from an input stream. It is a member function of the istream class, which is a base class for all input streams. The seekg() function takes two parameters:

streampos pos: This is the new position in the stream buffer.
ios_base::seekdir dir: This is a value that specifies the direction in which to seek. It can be one of the following values:
ios_base::beg: This seeks to the beginning of the stream.
ios_base::cur: This seeks to the current position in the stream.
ios_base::end: This seeks to the end of the stream.
The seekg() function returns a reference to the istream object on which it was called. This allows you to chain calls to seekg() and other member functions of the istream class.

Here is an example of how to use the seekg() function to seek to the beginning of a file:

C++
#include <iostream>
#include <fstream>

int main() {
  std::ifstream file("myfile.txt");

  if (file.is_open()) {
    // Seek to the beginning of the file
    file.seekg(0, std::ios_base::beg);

    // Read the contents of the file
    std::string line;
    while (getline(file, line)) {
      std::cout << line << std::endl;
    }

    file.close();
  } else {
    std::cerr << "Unable to open file" << std::endl;
  }

  return 0;
}
Use code with caution. Learn more
This code will open the file myfile.txt, seek to the beginning of the file, and then read the contents of the file line by line.

The seekg() function is a powerful tool that can be used to manipulate input streams in a variety of ways. It is often used in conjunction with other functions, such as tellg() and getline(), to implement more complex file processing tasks.
How to prompt for the updated data should be obvious. When the boolean Modified flag is true, then we want to write out the updated data to the file. If this flag is false then there is no sense in wasting the time to write unchanged data to the file. The basic idea is that we have to move the file position pointer, the one used to keep track of where to write to the file, to the correct location. Then we write out the record using our familiar write function that we always use with our binary files. Some form of "seek" command is used to move the file position pointer. If we want to use the pointer that keeps track of where to write, we must use the seekp version. The letter 'p' is a reminder that we plan to "put" some data into the file. (There is also a seekg version used when we want to move the file position pointer that keeps track of where we are going to read from a file. The 'g' is a reminder that we want to "get" data from the file.) The first parameter to either version of seek is the number of bytes to move into the file. The second parameter is a constant used to indicate where we are starting from. Typically one uses ios::beg which means that we will move a certain number of bytes into the file from the beginning.

Finally, let's look at the coding of the SearchFile function. This is the one that imitates a binary search within the file to try to find a record containing the first and last names in the Employee record. Of course, we already have an EmpCompare function to tell us if two records have names that match. In such a case EmpCompare returns 0. It returns a -1 if the employee record given as its first parameter is alphabetically less than the employee record given as its second parameter. See employee.cpp if you wish to look at the details.


bool SearchFile(fstream & EmpFile, EmployeeType & Employee, long & Location)
   {
   EmployeeType EmployeeTemp;
   bool Found;
   int CmpResult;
   long Mid, Low, High, RecordSize;

   Found = false;
   Low = 0L;
   // Go to the end of the file:
   EmpFile.seekg(0L, ios::end);
   RecordSize = sizeof(EmployeeTemp);
   // Find the number of records and subtract 1 to get high index:
   High = EmpFile.tellg() / RecordSize - 1L;

   while ((! Found) && (Low <= High))
      {
      Mid = (Low + High) / 2;
      Location = Mid * RecordSize;
      EmpFile.seekg(Location, ios::beg);
      EmpFile.read(reinterpret_cast <char *> (&EmployeeTemp), RecordSize);
      CmpResult = EmpCompare(Employee, EmployeeTemp);

      if (CmpResult == 0)
         {
         Employee = EmployeeTemp;
         Found = true;
         }
      else if (CmpResult < 0)
         High = Mid - 1L;
      else
         Low = Mid + 1L;
      }

   return Found;
   }
In the code above, note that any variable used to keep track of a position within the file is a long. Once again this is because the number of bytes in a file can be huge, so that an ordinary int might overflow. Variables Low, High, and Mid are used to hold record numbers, where the file's records are numbered 0, 1, 2, etc. Even these numbers could get to be rather large, so we use type long for them. Note that variable Low starts at value 0L, where the L is used to indicate a constant of type long.

Finding the proper location to use for variable High is somewhat tricky. We use seekg with the ios::end constant,